Rename: rename triptable to diarylist #19
